+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Alpha 3.9.3
+* Critical Hits now deal 25% more damage instead of 50%
 * Fixed timber getting more trees than expected
 
 ## Alpha 3.9.2